requirements.txt
tiktoken==0.12.0
python-dotenv==1.2.1
openai==2.14.0
scikit-learn==1.8.0
numpy==2.4.0